// features/rabbitmq/types.ts
export type ConsumerNodeData = {
  id: string;
  user: string;
  consumerTag: string;
  count: number;
  activeCount: number;
};